pyproject.toml
[project]
name = "smart-student-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "load-dotenv>=0.1.0",
    "openai-agents>=0.2.0",
]
